requirements.txt
ansible
dnspython